requirements.txt
google-generativeai==0.7.2
langchain==0.2.16
langchain-google-genai==1.0.10
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
beautifulsoup4==4.12.2
markdown==3.5.1